feature : ajout de la base da la marche continue (non fonctionnel)

+++ Assets/Scripts/CameraBehaviour.cs
@@ -13,8 +13,8 @@ public class CameraBehaviour : MonoBehaviour
 
     void LateUpdate()
     {
-        // Récupérer le score actuel du joueur
-        int playerScore = movement.GetScore();
+        // Récupérer le score actuel du joueur depuis UserExperienceMetronome
+        float playerScore = movement.GetScore();
 
         // Mapper le score du joueur à une position horizontale inversée
         // Si le score est minimum, l'offset est maxRightOffset (la caméra est à droite)
